content(about): prevent repeat pinky frags
Looked a bit unintentional. This is more authored. 6 clicks was
chosen as the optimally comedic number. May add an even deeper
easter egg (like after a delay, or maybe if you click 10 times
an arch-vile comes by and revives the pinky (resetting click counter)).

Fun with JavaScript! Is it any wonder this stuff took over the web??