fix(api): change API path
This brings it in line with the package and dir names, and I think it
makes sense.

Yes I know I'm so naughty, no versioning or anything--just changing the
path. This is of course a breaking change, but as of writing this it's
occurring on a branch so I'm not really concerned about someone (me).

But still

BREAKING CHANGE: GET /interest/{slug} ~> /interests/{slug}